backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.25.2
pydantic==2.5.0
python-dotenv==1.0.0
supabase==2.3.4
groq==0.4.1
web3==6.11.3
xhtml2pdf==0.2.11
reportlab==4.0.7
shap==0.43.0
requests==2.31.0
aiofiles==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4